requirements.txt
accelerate==0.29.3
bitsandbytes==0.43.1
clip @ git+https://github.com/openai/CLIP.git  
datasets==3.1.0
diffusers==0.24.0
einops==0.8.0
facexlib==0.3.0
ftfy==6.2.0
huggingface-hub==0.25.0
imageio==2.36.0
insightface==0.7.3
Jinja2==3.1.3
matplotlib==3.9.2
numpy==1.26.4
onnx==1.17.0
packaging==24.2
pandas==2.2.3
pillow==10.3.0
scipy==1.14.1
safetensors==0.4.3
tensorboard==2.16.2
timm==1.0.11
torch==2.2.2
torchvision==0.17.2
wandb==0.16.6
transformers==4.46.2
onnxruntime==1.20.0